[PAGOPA-1482] test: refactoring unit test
