+++ src/test/java/it/gov/pagopa/gpd/upload/functions/util/MapUtilsTest.java
@@ -1,0 +1,17 @@
+package it.gov.pagopa.gpd.upload.functions.util;
+
+import it.gov.pagopa.gpd.upload.model.UploadReport;
+import it.gov.pagopa.gpd.upload.util.MapUtils;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import static it.gov.pagopa.gpd.upload.functions.utils.TestUtil.getMockStatus;
+
+public class MapUtilsTest {
+
+    @Test
+    void runOK() {
+        UploadReport report = MapUtils.convert(getMockStatus());
+        Assertions.assertNotNull(report);
+    }
+}
